UIFlow 3.3 compatibility changes
