--- boards/m5stack-cardputer-adv/platformio.ini
@@ -1,77 +1,0 @@
-; PlatformIO Project Configuration File
-;
-;   Build options: build flags, source filter
-;   Upload options: custom upload port, speed and extra flags
-;   Library options: dependencies, extra library storages
-;   Advanced options: extra scripting
-;
-; Please visit documentation for the other options and examples
-; https://docs.platformio.org/page/projectconf.html
-
-[env:m5stack-cardputer-adv]
-board = m5stack-cardputer
-extra_scripts =
-	${env.extra_scripts}
-	pre:support_files/nvs_script.py
-board_build.partitions = custom_8Mb.csv
-build_src_filter =${env.build_src_filter} +<../boards/m5stack-cardputer-adv>
-build_flags =
-	${env.build_flags}
-	-Iboards/m5stack-cardputer-adv
-	-DCORE_DEBUG_LEVEL=0
-	-DARDUINO_USB_CDC_ON_BOOT=1
-	-DCARDPUTER=1
-	-DCARDPUTER_ADV=1
-	-DEEPROMSIZE=128
-	-DTCA8418_INT_PIN=11
-	-DTCA8418_I2C_ADDR=0x34
-	-DTCA8418_SDA_PIN=8
-	-DTCA8418_SCL_PIN=9
-	-D JSON_SOURCE_PATH='"https://raw.githubusercontent.com/bmorcelli/M5Stack-json-fw/main/v2/cardputer.json"'
-
-	;-DPART_04MB =0
-	-DPART_08MB=1
-	;-DPART_16MB=0
-
-	-DROTATION=1
-	-DHAS_BTN=1
-	-DBTN_ALIAS='"Ok"'
-	-DBTN_PIN=0
-
-	-DFP=1
-	-DFM=2
-	-DFG=3
-	-DBAT_PIN=10
-	-DLED=21
-	-DLED_ON=HIGH
-	-DBACKLIGHT=38
-	-DMINBRIGHT=165
-
-
-
-	-DST7789_DRIVER=1
-	-DTFT_WIDTH=135
-	-DTFT_HEIGHT=240
-	-DTFT_BL=38
-	-DTFT_RST=33
-	-DTFT_DC=34
-	-DTFT_MOSI=35
-	-DTFT_SCLK=36
-	-DTFT_CS=37
-
-	-D GFX_BL=38
-	-D TFT_IPS=1
-	-D TFT_COL_OFS1=52
-	-D TFT_ROW_OFS1=40
-	-D TFT_COL_OFS2=53
-	-D TFT_ROW_OFS2=40
-
-	-DSDCARD_SPI_BUS=FSPI
-	-DSDCARD_CS=12
-	-DSDCARD_SCK=40
-	-DSDCARD_MISO=39
-	-DSDCARD_MOSI=14
-
-lib_deps =
-	${env.lib_deps}
-	adafruit/Adafruit TCA8418 @ ^1.0.1
