+++ boards/m5stack-cardputer/platformio.ini
@@ -41,7 +41,10 @@ build_flags =
 	-DBACKLIGHT=38
 	-DMINBRIGHT=165
 
-
+	-DTCA8418_INT_PIN=11
+	-DTCA8418_I2C_ADDR=0x34
+	-DTCA8418_SDA_PIN=8
+	-DTCA8418_SCL_PIN=9
 
 	-DST7789_DRIVER=1
 	-DTFT_WIDTH=135
@@ -68,3 +71,4 @@ build_flags =
 
 lib_deps =
 	${env.lib_deps}
+	adafruit/Adafruit TCA8418 @ ^1.0.1
