+++ boards/marauder-mini/interface.cpp
@@ -1,11 +1,6 @@
 #include "powerSave.h"
 #include <interface.h>
 
-#include <driver/adc.h>
-#include <esp_adc_cal.h>
-#include <soc/adc_channel.h>
-#include <soc/soc_caps.h>
-
 /***************************************************************************************
 ** Function name: _setup_gpio()
 ** Location: main.cpp
