+++ boards/marauder-v4og/interface.cpp
@@ -2,10 +2,6 @@
 #include <interface.h>
 
 #include <CYD28_TouchscreenR.h>
-#include <driver/adc.h>
-#include <esp_adc_cal.h>
-#include <soc/adc_channel.h>
-#include <soc/soc_caps.h>
 
 CYD28_TouchR touch(320, 240);
