+++ boards/smoochiee-board/interface.cpp
@@ -1,10 +1,5 @@
 #include "powerSave.h"
 
-#include <driver/adc.h>
-#include <esp_adc_cal.h>
-#include <soc/adc_channel.h>
-#include <soc/soc_caps.h>
-
 /***************************************************************************************
 ** Function name: _setup_gpio()
 ** Location: main.cpp
