+++ custom_8Mb.csv
@@ -1,23 +1,23 @@
 # Default, , , , ,
 # Name,   Type, SubType, Offset,   Size, Flags
 nvs,      data, nvs,     0x9000,   0x6000,
-app0,     app,  test,    0x10000,  0x180000,
-app1,     app,  ota_0,   0x190000, 0x4E0000,
+app0,     app,  test,    0x10000,  0x160000,
+app1,     app,  ota_0,   0x170000, 0x4F0000,
 vfs,      data, fat,     0x670000, 0x80000,
 spiffs,   data, spiffs,  0x6F0000, 0x100000,
-coredump, data, coredump,0x7F0000, 0x10000, 
+coredump, data, coredump,0x7F0000, 0x10000,
 #
 # Doom, , , , ,
 # Name,   Type, SubType, Offset,   Size, Flags
 #nvs,      data, nvs,     0x9000,   0x6000,
-#app0,     app,  test,    0x10000,  0x180000,
-#app1,     app,  ota_0,   0x190000, 0x660000,
-#coredump, data, coredump,0x7F0000, 0x10000, 
+#app0,     app,  test,    0x10000,  0x160000,
+#app1,     app,  ota_0,   0x170000, 0x680000,
+#coredump, data, coredump,0x7F0000, 0x10000,
 #
 # UiFlow, , , , ,
 # Name,   Type, SubType, Offset,   Size, Flags
 #nvs,      data, nvs,     0x9000,   0x6000,
-#app0,     app,  test,    0x10000,  0x170000,
-#app1,     app,  ota_0,   0x180000, 0x4B0000,
-#sys,      data, fat,     0x630000, 0x100000,
-#vfs,      data, fat,     0x730000, 0xD0000,
+#app0,     app,  test,    0x10000,  0x160000,
+#app1,     app,  ota_0,   0x170000, 0x4E0000,
+#sys,      data, fat,     0x650000, 0x100000,
+#vfs,      data, fat,     0x750000, 0xB0000,
