+++ platformio.ini
@@ -62,6 +62,7 @@ board_upload.offset_address = 0x10000
 extra_scripts =
 	pre:support_files/prep_web_files.py
 	pre:support_files/upload_patch.py
+	post:support_files/merge.py
 build_flags =
 	-DLAUNCHER='"dev"'
 	-DMAXFILES=256
