+++ support_files/merge.py
@@ -63,9 +63,9 @@ def _merge_bins_callback(target, source, env):
 
     # Initialize cmd_parts as a list
     cmd_parts = [
-        "pio pkg exec -p \"tool-esptoolpy\" -- esptool.py",
+        "pio pkg exec -p \"tool-esptoolpy\" -- esptool",
         "--chip", chip_arg,
-        "merge_bin",
+        "merge-bin",
         "--output", q(out_bin),
         hex(boot_offset), q(boot_bin),
         hex(PART_TABLE_OFFSET), q(part_bin),
